Iterative works fine , attempting to fix reporting with recursive. Cleaned up some code also and now have two functions to calculate teh standard deviation as well as the average

+++ src/main/java/SortInterface.java
@@ -3,7 +3,7 @@
  */
 public interface SortInterface {
 
-    int recursiveSort(int[] list, int start);
+    long recursiveSort(int[] list, int start);
     void iterativeSort(int[] list);
     long getCount();
     long getTime();
